% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalo-package.R
\docType{package}
\name{dalo-package}
\alias{dalo}
\alias{dalo-package}
\title{dalo: Data Download Helper Tools}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Syunsuke Fukuda \email{ok.xmonad@gmail.com}

}
\keyword{internal}
