% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_by_urls.R
\name{dalo_dl_by_url_with_name}
\alias{dalo_dl_by_url_with_name}
\title{Download A file with a favorit name}
\usage{
dalo_dl_by_url_with_name(
  url,
  dest_dir = "./",
  filename,
  check = TRUE,
  make_dir = FALSE
)
}
\arguments{
\item{url}{single url string}

\item{dest_dir}{directory which files get into}

\item{filename}{which you'd like to name}

\item{check}{if TRUE, don't download exist files}

\item{make_dir}{if TRUE, make a directory}
}
\description{
Download A file with a favorit name
}
