% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlhelper.R
\name{dalo_download_by_urls}
\alias{dalo_download_by_urls}
\title{Download Files Easily}
\usage{
dalo_download_by_urls(urls, dest_dir = "./", check = TRUE)
}
\arguments{
\item{urls}{url string vector}

\item{dest_dir}{directory which files get into}

\item{check}{if TRUE, check downloaded files}
}
\description{
Download Files Easily
}
