% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_link.R
\name{dalo_fetch_links}
\alias{dalo_fetch_links}
\title{fetch link urls from web page}
\usage{
dalo_fetch_links(
  url,
  selector = "",
  filename_pattern = ".*",
  encoding = "utf-8",
  prefix = ""
)
}
\arguments{
\item{url}{WebPage url or html file path}

\item{selector}{CSS selector}

\item{filename_pattern}{Regular expressions}

\item{encoding}{utf-8 or shift-jis and so on}

\item{prefix}{prefix strings}
}
\value{
vector of links
}
\description{
fetch link urls from web page
}
