% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickup_from_zipfile.R
\name{dalo_pickup_from_zipfile}
\alias{dalo_pickup_from_zipfile}
\title{PickUp some files From Zipfile}
\usage{
dalo_pickup_from_zipfile(zipfiles, pattern = "", dest_dir = ".")
}
\arguments{
\item{zipfiles}{vector of zipfiles path}

\item{pattern}{pattern of files which you want}

\item{dest_dir}{destination where pickuped file goes}
}
\description{
PickUp some files From Zipfile
}
