% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_by_urls.R
\name{dalo_download_by_urls}
\alias{dalo_download_by_urls}
\title{Download Files Easily}
\usage{
dalo_download_by_urls(urls, dest_dir = "./", check = TRUE, make_dir = FALSE)
}
\arguments{
\item{urls}{url string vector}

\item{dest_dir}{directory which files get into}

\item{check}{if TRUE, don't download exist files}

\item{make_dir}{if TRUE, make a directory}
}
\description{
Download Files Easily
}
